% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARC-model.R
\name{ARC}
\alias{ARC}
\alias{plot.ARC}
\alias{ARC}
\alias{ARC}
\alias{ARC.default}
\alias{ARC}
\alias{ARC.PATH}
\title{ARC model}
\usage{
\method{plot}{ARC}(x, ..., lwd = 2L)

ARC(x, ...)

\method{ARC}{default}(x, ...)

\method{ARC}{PATH}(x, ...)
}
\arguments{
\item{x}{input model}

\item{...}{passed to methods}
}
\value{
ARC model
}
\description{
Arcs are paths within the line segment graph between nodes.
}
\details{
Nodes are the vertices where three or more arcs meet. 
An arc can exist without including any nodes, a path that has no neighbouring relationship with another path.

This is _not_ the same terminology as used by other systems. The `arc_link_vertex` mapping is inherently ordered, 
but we don't consider order of arcs. Duplicated arcs (i.e. complementary turns around neighbouring polygons) are 
not kept. The `object_link_arc` mapping records which arc belongs to the objects, so can be reconstructed within 
objects by tracing `arc_link_vertex` start and end point identity.
}
\examples{
a <- ARC(minimal_mesh)
sc_arc(a)
sc_arc(minimal_mesh)
}
