% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives_1D.r
\name{arc_node}
\alias{arc_node}
\alias{arc_node}
\alias{arc_node.sf}
\alias{arc_node}
\alias{arc_node.PRIMITIVE}
\title{Arc-node topology.}
\usage{
arc_node(x, ...)

\method{arc_node}{sf}(x, ...)

\method{arc_node}{PRIMITIVE}(x, ...)
}
\arguments{
\item{x}{input model}

\item{...}{arguments to methods}
}
\value{
`tbl_df` of the node coordinates
}
\description{
Arcs are unbranched paths within the line segment graph. Nodes are the vertices where three or more arcs meet.
}
\examples{
x <- sf::st_read(system.file("extdata/file.geojson", package= "sc"))
arc_node(x)  ## get the nodes
## now get the arcs (should the functions be called arc() and node()?)
}
