% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UGRID-convert.R
\name{as_mesh3d}
\alias{as_mesh3d}
\title{rgl mesh3d, quads kept as quads}
\usage{
as_mesh3d(x, z = 0)
}
\arguments{
\item{x}{a UGRID model.}

\item{z}{optional node heights.}
}
\description{
rgl mesh3d, quads kept as quads
}
