% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINARY.R
\name{BINARY}
\alias{BINARY}
\title{Binary form}
\usage{
BINARY(x, ...)
}
\arguments{
\item{x}{object (sf, sp, anything undertood by silicate/gibble)}

\item{...}{ignored}
}
\value{
binary form
}
\description{
Build a purely edge-based form from path-based data.
}
\details{
Binary is a two-table form of \code{object} and \code{vertex}. On object the \code{vertex_} is a list column,
with \code{.vx0} and \code{.vx1}. These are the row numbers of \code{vertex}.
}
\examples{
plot(BINARY(minimal_mesh))
#library(trip)
#plot(BINARY(walrus818))

#b <- BINARY(rnaturalearth::ne_countries(returnclass = "sp"))
#plot(b)
# we can subset object trivially, with no need to relabel
# because we aren't changing the vertex pool
# b$object <- dplyr::filter(b$object, nchar(sovereignt) > 13)
# plot(b)
}
