% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRANCH-model.r
\name{BRANCH}
\alias{BRANCH}
\alias{BRANCH.default}
\title{Generate a BRANCH model.}
\usage{
BRANCH(x, ...)

\method{BRANCH}{default}(x, ...)
}
\arguments{
\item{x}{input model}

\item{...}{arguments passed to methods}
}
\description{
A BRANCH model is a direct translation of the a simple features-alike
model to normal form.
}
\examples{
library(sf)
sf_dataset <- st_sf(geometry = st_sfc(sfzoo[[2]]), a = 1)
BRANCH(sf_dataset)
}
\seealso{
`sc_branch`, `sc_coord`
}

