% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-builder.R
\name{build_sf}
\alias{build_sf}
\title{a pattern for building sf objects from
\itemize{
\item a gibble, the map of the paths
\item the coordinates
the map is an encoding of the structural properties of the geometry
}}
\usage{
build_sf(x, ...)
}
\description{
currently only XY is supported
}
