% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dplyr-methods}
\alias{dplyr-methods}
\alias{filter.SC}
\title{Dplyr methods for silicate objects}
\usage{
filter.SC(x, ...)
}
\arguments{
\item{...}{other arguments}

\item{.data}{data object of class \link{silicate-models}}

\item{.dots}{see corresponding function in package \code{dplyr}}
}
\description{
dplyr methods for silicate objects, currently only \code{dplyr::filter} for \code{SC} is available.
}
\details{
Apply expressions as if used on the \code{object} table. See \code{sc_object(x)} for that form.

Currently all the vertices are still kept, so the model (and any plots) include the filtered edges as well
as undifferentiaed points. This is likely to change ...
}
\examples{
library(dplyr)
sc <- SC(inlandwaters)
plot(filter(sc, Province == "Tasmania"))

plot(filter(SC(minimal_mesh), a == 2))
}
