% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silicate-package.R
\docType{data}
\name{ear_gc}
\alias{ear_gc}
\title{Geometry collection of triangles}
\description{
A 'sfc_GEOMETRYCOLLECTION' of four-cornered triangles ('POLYGON')
created by ear cutting the North Carolina polygon from sf.
}
\examples{
TRI0(ear_gc)
}
