% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earcut.R
\name{earcut}
\alias{earcut}
\alias{earcut.PATH}
\alias{earcut}
\alias{earcut.sf}
\alias{earcut}
\alias{earcut.sfc}
\title{Ear cut triangulation}
\usage{
\method{earcut}{PATH}(x, holes = NULL, ...)

\method{earcut}{sf}(x, ...)

\method{earcut}{sfc}(x, ...)
}
\arguments{
\item{x}{PATH object to triangulate}

\item{holes}{ignored by the PATH method}

\item{...}{}
}
\description{
Ear cutting triangulation for polygons.
}
\details{
This function extends the \code{earcut} generic from the package \code{rearcut}.
}
\examples{
earcut(PATH(minimal_mesh))
}
