% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact-indexes.R
\name{expand_indexes}
\alias{expand_indexes}
\alias{expand_indexes}
\alias{expand_indexes.compact_SC}
\alias{expand_indexes}
\alias{expand_indexes.compact_PATH}
\title{Expand indexes from structural compaction}
\usage{
expand_indexes(x, ...)

\method{expand_indexes}{compact_SC}(x, ...)

\method{expand_indexes}{compact_PATH}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
non-compact form, i.e. SC or PATH
}
\description{
Expand indexes from structural compaction
}
\examples{
plot(SC(expand_indexes(compact_indexes(PATH(minimal_mesh)))))
}
