% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UGRID-verbs.R
\name{mesh_dims}
\alias{mesh_dims}
\title{Keyed dimensions of a UGRID model}
\usage{
mesh_dims(x)
}
\arguments{
\item{x}{a UGRID model.}
}
\value{
data.frame: location, the netCDF dimension name it is keyed
to, and its length.
}
\description{
Keyed dimensions of a UGRID model
}
