% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silicate-package.R
\docType{data}
\name{minimal_mesh}
\alias{minimal_mesh}
\title{Minimal mesh.}
\description{
The simplest pairing of two polygons with one shared edge. One polygon
contains a hole and a concavity, the other is a simply convex. This is
composed of four "arcs", one around each polygon, one for the shared edge, and
one for the isolated hole. There are two nodes, the endpoints of the single shared edge.
}
\examples{
arc <- ARC(minimal_mesh)
plot(arc)
sc_arc(arc)
sc_node(arc)
}
