% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silicate-package.R
\docType{data}
\name{mmesh}
\alias{mmesh}
\title{Deprecated data set.}
\description{
This data set is in legacy format and will be removed.
A couple of polygons with a single shared edge between them, in
PRIMITIVE form.
}
