% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives_1D.r
\name{NARC}
\alias{NARC}
\alias{NARC}
\alias{NARC.PRIMITIVE}
\title{NARC model, arc-node topology.}
\usage{
NARC(x, ...)

\method{NARC}{PRIMITIVE}(x, ...)
}
\arguments{
\item{x}{input model}

\item{...}{arguments to methods}
}
\value{
`tbl_df` of the node coordinates
}
\description{
Arcs are unbranched paths within the line segment graph. Nodes are the vertices where three or more arcs meet.
}
