% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATH0-model.R
\name{PATH0}
\alias{PATH0}
\alias{PATH0.default}
\title{Path model in structural form}
\usage{
PATH0(x, ...)

\method{PATH0}{default}(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Path model in structural form
}
