% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINARY.R
\name{plot.binary}
\alias{plot.binary}
\title{Plot BINARY}
\usage{
\method{plot}{binary}(x, ...)
}
\arguments{
\item{x}{BINARY object}

\item{...}{arguments  passed to \code{graphics::segments} (col is ignored)}
}
\description{
Plot a BINARY model, segments coloured by object.
}
