% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC0-model.R
\name{plot.SC0}
\alias{plot.SC0}
\title{Plot silicate}
\usage{
\method{plot}{SC0}(x, ..., add = FALSE)
}
\arguments{
\item{x}{SC0 object}

\item{...}{arguments  passed to \code{graphics::segments} (col is ignored)}
}
\description{
Plot a SC0 model, primitives coloured by object.
}
