% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silicate-package.R
\docType{data}
\name{polymesh}
\alias{polymesh}
\title{Polygonal mesh}
\description{
A simple set of \code{sf} neighbouring polygons, with redundant vertices created
from polygonizing a raster.
}
\examples{
arc <- ARC(polymesh)
plot(arc)
sc <- SC(polymesh)
plot(sc)
}
