% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIM.R
\name{PRIMITIVE}
\alias{PRIMITIVE}
\alias{PRIMITIVE}
\alias{PRIMITIVE.PATH}
\alias{PRIMITIVE}
\alias{PRIMITIVE.default}
\title{Generate a PRIMITIVES model.}
\usage{
PRIMITIVE(x, ...)

\method{PRIMITIVE}{PATH}(x, ...)

\method{PRIMITIVE}{default}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments passed to methods}
}
\description{
A PRIMITIVES model is a decomposition of spatial forms to primitives.
}
