% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC-model.R
\name{SC}
\alias{SC}
\alias{SC}
\alias{plot.SC}
\title{The universal model}
\usage{
SC(x, ...)

\method{plot}{SC}(x, ..., vars = NULL)
}
\arguments{
\item{x}{input model}

\item{...}{arguments passed to methods}

\item{vars}{variables to plot}
}
\description{
The universal model `SC` is coordinates and binary relations between
pairs of coordinates. This is purely an edge (or segment) model, with all 
higher level structures recorded as groupings of edges.
}
