% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC0-model.R
\name{SC0}
\alias{SC0}
\alias{SC0}
\alias{SC0.default}
\title{core silicate}
\usage{
SC0(x, ...)

\method{SC0}{default}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{reserved for methods}
}
\value{
SC0
}
\description{
See https://github.com/hypertidy/silicore for now.
}
\examples{
SC0(silicate::minimal_mesh)
}
