% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives_1D.r
\name{sc_arc}
\alias{sc_arc}
\alias{sc_arc}
\alias{sc_arc.default}
\alias{sc_arc}
\alias{sc_arc.PRIMITIVE}
\title{Arcs for arc-node topology.}
\usage{
sc_arc(x, ...)

\method{sc_arc}{default}(x, ...)

\method{sc_arc}{PRIMITIVE}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments for methods}
}
\description{
So-called "arcs" are unclosed paths that end in nodes, vertices shared by other arcs.
}
