% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.r
\name{sc_branch}
\alias{sc_branch}
\alias{sc_branch.GEOMETRYCOLLECTION}
\alias{sc_branch.LINESTRING}
\alias{sc_branch.MULTILINESTRING}
\alias{sc_branch.MULTIPOINT}
\alias{sc_branch.MULTIPOLYGON}
\alias{sc_branch.POINT}
\alias{sc_branch.POLYGON}
\alias{sc_branch.sf}
\alias{sc_branch.sfc}
\title{Branch decomposition}
\usage{
sc_branch(x, ...)

\method{sc_branch}{sf}(x, ...)

\method{sc_branch}{sfc}(x, ids = NULL, ...)

\method{sc_branch}{MULTIPOLYGON}(x, ...)

\method{sc_branch}{POLYGON}(x, ...)

\method{sc_branch}{LINESTRING}(x, ...)

\method{sc_branch}{MULTILINESTRING}(x, ...)

\method{sc_branch}{POINT}(x, ...)

\method{sc_branch}{MULTIPOINT}(x, ...)

\method{sc_branch}{GEOMETRYCOLLECTION}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments passed to methods}

\item{ids}{object id, one for each object in the `sfc`}
}
\description{
Start in the middle, and build the 'branch-link-vertex' table.
}
\examples{
data("sfzoo")
lapply(sfzoo, sc_branch)
sc_branch(sf::st_sfc(sfzoo))
sc_branch(sf::st_sfc(sfzoo))

sc_branch(sfzoo$multipolygon)
sc_branch(sfzoo$polygon)
sc_branch(sfzoo$linestring)
sc_branch(sfzoo$multilinestring)
sc_branch(sfzoo$point)
sc_branch(sfzoo$multipoint)
sc_branch(sfzoo$multipoint)
}
\seealso{
`sc_coord` for the coordinates part of the model, `sc_object` for 
the features, and `BRANCH` for the full model.
}

