% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sc_colours}
\alias{sc_colours}
\title{Silicate colours}
\usage{
sc_colours(x = 16, ..., viridis = FALSE)
}
\arguments{
\item{x}{number of colours to generate}

\item{...}{currently ignored}

\item{viridis}{use viridis, TRUE or FALSE}
}
\value{
vector of colours
}
\description{
Simple set of colours for discrete palette.
}
\examples{
sc_colours(10)
}
