% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{sc_object.RasterLayer}
\alias{sc_object.RasterLayer}
\title{object-value for cell-based raster (discrete pixels)
(continuous pixels store values with coordinates)}
\usage{
\method{sc_object}{RasterLayer}(x, ...)
}
\description{
object-value for cell-based raster (discrete pixels)
(continuous pixels store values with coordinates)
}
