% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.r
\name{sc_object}
\alias{sc_object}
\alias{sc_object}
\alias{sc_object.default}
\title{Objects, features}
\usage{
sc_object(x, ...)

\method{sc_object}{default}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments passed to methods}
}
\description{
The objects are the front end entities, the usual "GIS contract" objects, 
or features.
}
\seealso{
`sc_coord` for the coordinates part of the model, `sc_path` for 
the central part of the model, and `PATH` for the full model.
}
