% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives_1D.r
\name{sc_primitive}
\alias{sc_primitive}
\title{Given a `BRANCH`` model decompose to 1-dimensional primitives.}
\usage{
sc_primitive(x, ...)
}
\description{
Given a `BRANCH`` model decompose to 1-dimensional primitives.
}
\examples{

sf_obj <- inlandwaters[5, ]
obj <- BRANCH(sf_obj)
prim <- sc_primitive(obj)
sf_pslg <- PRIMITIVE(sf_obj)
}

