% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_segment.R
\name{sc_segment}
\alias{sc_segment}
\alias{sc_segment.default}
\alias{sc_segment.PATH}
\title{Given a `PATH`` model decompose to 1-dimensional primitives (or 0-dimensional).}
\usage{
sc_segment(x, ...)

\method{sc_segment}{default}(x, ...)

\method{sc_segment}{PATH}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments passed to methods}
}
\value{
data frame of the segments, each occurence of an edge and its order
}
\description{
Given a `PATH`` model decompose to 1-dimensional primitives (or 0-dimensional).
}
\examples{
sc_segment(SC(minimal_mesh))
}
