% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_triangle.R
\name{sc_triangle}
\alias{sc_triangle}
\alias{sc_triangle.default}
\alias{sc_triangle.TRI0}
\alias{sc_triangle.TRI}
\alias{sc_triangle.mesh3d}
\title{Silicate methods}
\usage{
sc_triangle(x, ...)

\method{sc_triangle}{default}(x, ...)

\method{sc_triangle}{TRI0}(x, ...)

\method{sc_triangle}{TRI}(x, ...)

\method{sc_triangle}{mesh3d}(x, ...)
}
\arguments{
\item{prefer_triangles}{give triangles in preference to quads, if quads also present}
}
\description{
Silicate methods
}
