% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{sc_uid}
\alias{sc_uid}
\title{Unique labels}
\usage{
sc_uid(x, ...)
}
\arguments{
\item{x}{number of unique IDs to generate}

\item{...}{reserved for future use}
}
\description{
Find unique labels for entities, or create them
if not present.
}
\details{
See `ids` package for `random_id` when they are created.
}
