% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_vertex.R
\name{sc_vertex}
\alias{sc_vertex}
\alias{sc_vertex}
\alias{sc_vertex.default}
\alias{sc_vertex}
\alias{sc_vertex.SC}
\alias{sc_vertex}
\alias{sc_vertex.ARC}
\alias{sc_vertex}
\alias{sc_vertex.TRI}
\alias{sc_vertex}
\alias{sc_vertex.PATH}
\title{Extract unique vertices}
\usage{
sc_vertex(x, ...)

\method{sc_vertex}{default}(x, ...)

\method{sc_vertex}{SC}(x, ...)

\method{sc_vertex}{ARC}(x, ...)

\method{sc_vertex}{TRI}(x, ...)

\method{sc_vertex}{PATH}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{passed to methods}
}
\description{
Extract unique vertices
}
