% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-package.r
\docType{package}
\name{sfzoo}
\alias{sfgc}
\alias{sfzoo}
\alias{sfzoo-package}
\title{Simple features zoo.}
\description{
A list with basic examples of each type of simple feature.
}
\details{
\code{sfgc} is a GEOMETRYCOLLECTION, separate from the \code{sfzoo}.

The gc contains the zoo.  
 @examples 
 sf::st_sfc(sfzoo)
 sf::st_sfc(sfgc)
}

