% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives_1D.r
\name{st_as_sf.PRIMITIVE}
\alias{st_as_sf.PRIMITIVE}
\title{Recompose `sf` simple features from `PRIMITIVE`` models.}
\usage{
\method{st_as_sf}{PRIMITIVE}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments passed to methods}
}
\description{
Recompose `sf` simple features from `PRIMITIVE`` models.
}
\examples{
prim <- PRIMITIVE(inlandwaters)
library(sf)
plot(st_as_sf(prim))
}

