% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_to_sf.R
\name{to_sf.sc}
\alias{to_sf.sc}
\alias{to_sf}
\title{Build sf from silicate}
\usage{
\method{to_sf}{sc}(x, sfc_type = NULL, geom_type = NULL)
}
\arguments{
\item{x}{silicate model}
}
\value{
sf dataframe
}
\description{
Build sf
}
\examples{
to_sf(PATH(minimal_mesh))
}
