% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRI-model.R, R/plot.R
\name{TRI}
\alias{TRI}
\alias{plot.TRI}
\title{TRI model, triangulations}
\usage{
TRI(x, ...)

\method{plot}{TRI}(x, ..., add = FALSE)
}
\arguments{
\item{x}{object understood by silicate (sf, sp, a silicate model, etc.)}

\item{...}{current unused}

\item{add}{logical create  new plot (default), or add to existing}
}
\value{
TRI model
}
\description{
TRI model, triangulations
}
