% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRI-model.R
\name{triangulate.SC}
\alias{triangulate.SC}
\title{Triangulation}
\usage{
triangulate.SC(x, ...)
}
\description{
Triangulation
}
\examples{
dd <- minimal_mesh
#dd <- nc
#dd <- rnaturalearth::ne_countries(scale = 50, returnclass = "sf")
x <- SC(dd)
plot(x)
xt <- triangulate.SC(x, D = TRUE)
plot(xt)
}
