% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UGRID-convert.R
\name{ugrid_polygons}
\alias{ugrid_polygons}
\title{Faces as polygons}
\usage{
ugrid_polygons(x)
}
\arguments{
\item{x}{a UGRID model.}
}
\value{
a wk wkb vector, one polygon per face.
}
\description{
Faces as polygons
}
