% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UGRID-io.R
\name{ugrid_read_array}
\alias{ugrid_read_array}
\title{Read one array as a keyed array}
\usage{
ugrid_read_array(dsn, variable, x, ..., dim_map = NULL)
}
\arguments{
\item{dsn}{netCDF file.}

\item{variable}{variable name.}

\item{x}{a UGRID model.}

\item{...}{other dimensions to slice, as name = 1-based index, e.g.
\code{time = 13}. Unsliced dimensions are read in full.}

\item{dim_map}{as in \code{\link[=ugrid_arrays]{ugrid_arrays()}}.}
}
\value{
a keyed array (the structure of \code{meshcore::keyed_array()})
keyed on the mesh dimension, with attribute "location", ready for
\code{meshcore::join_array()}.
}
\description{
Read one array as a keyed array
}
