% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo.R
\name{zoo_matrix}
\alias{zoo_matrix}
\title{The conversion matrix}
\usage{
zoo_matrix()
}
\value{
character matrix, rows from, columns to: "lossless", "lossy",
or "" for a missing route.
}
\description{
The conversion matrix
}
