% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRANCH-model.r
\name{BRANCH}
\alias{BRANCH}
\alias{BRANCH.default}
\title{Generate a BRANCH model.}
\usage{
BRANCH(x, ...)

\method{BRANCH}{default}(x, ...)
}
\arguments{
\item{x}{input model}

\item{...}{arguments passed to methods}
}
\description{
A BRANCH model is a direct translation of a simple features-alike
object to normal form. This is four tables with the three kinds of entities, 
"objects" (or "features"), "branches" (or "parts") and "vertices", and a table 
to link the one-to-many relation between branches and vertices.
}
\details{
In a data set with no parts touching their  neighbours, the only normalization of the vertices
will be the removal of the duplicated closing coordinate on any polygon ring, and on
any self-intersecting case within a single branch.
}
\examples{
library(sf)
sf_dataset <- st_sf(geometry = st_sfc(sfzoo[[2]]), a = 1)
BRANCH(sf_dataset)
}
\seealso{
`sc_branch`, `sc_coord`
}

