% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-builder.R
\name{build_sf}
\alias{build_sf}
\title{a pattern for building sf objects from 
- a gibble, the map of the paths
- the coordinates
the map is an encoding of the structural}
\usage{
build_sf(gm, coords_in, crs = NULL)
}
\description{
a pattern for building sf objects from 
- a gibble, the map of the paths
- the coordinates
the map is an encoding of the structural
}
