% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact-indexes.R
\name{compact_indexes}
\alias{compact_indexes}
\alias{compact_indexes}
\alias{compact_indexes.SC}
\alias{compact_indexes}
\alias{compact_indexes.PATH}
\title{Compact form, structural indexing}
\usage{
compact_indexes(x, ...)

\method{compact_indexes}{SC}(x, ...)

\method{compact_indexes}{PATH}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
a special "compact_*" form, of e.g.  SC, or PATH
}
\description{
Compact form, structural indexing
}
\examples{
expand_indexes(SC(minimal_mesh))
}
