% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earcut.R
\name{earcut}
\alias{earcut}
\alias{earcut.PATH}
\alias{earcut}
\alias{earcut.sf}
\alias{earcut}
\alias{earcut.sfc}
\title{Ear cut triangulation}
\usage{
earcut.PATH(x, holes = NULL, ...)

earcut.sf(x, ...)

earcut.sfc(x, ...)
}
\arguments{
\item{x}{PATH object to triangulate}

\item{holes}{ignored by the PATH method}

\item{...}{}
}
\description{
Ear cutting triangulation for polygons.
}
\details{
This function builds on \code{earcut} the package \code{decido}.
}
\examples{
earcut.PATH(PATH(minimal_mesh))
}
