% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-package.r
\docType{data}
\name{inlandwaters}
\alias{inlandwaters}
\title{Inland waters, for parts of Australia.}
\description{
The inland waters are lake and so on, presenting as holes
within the bounded regions of Australian provinces.
}
\details{
This is an extract from the old Manifold DVD. It is in `sf` format`.
}
