% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATH0-model.R
\name{PATH0}
\alias{PATH0}
\alias{PATH0.default}
\alias{PATH0.PATH0}
\title{Path model in structural form}
\usage{
PATH0(x, ...)

\method{PATH0}{default}(x, ...)

\method{PATH0}{PATH0}(x, ...)
}
\arguments{
\item{x}{an object understood by silicate}

\item{...}{ignored currently}
}
\value{
PATH0 model with tables 'object' and 'vertex'
}
\description{
Structural form requires only tables 'object' and 'vertex'.
}
\examples{
(p <- PATH0(minimal_mesh))

p$object$topology_
}
