% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.SC}
\alias{plot.SC}
\alias{plot.SC0}
\title{Plot silicate}
\usage{
\method{plot}{SC}(x, ..., add = FALSE)

\method{plot}{SC0}(x, ..., add = FALSE)
}
\arguments{
\item{x}{sc object}

\item{...}{arguments passed to lower level plotting functions}

\item{add}{if \code{TRUE} add to current plot}
}
\description{
Basic edge plot, all the standard base graphics facilities for line segments are available.
}
\details{
The 'col' argument is passed directly to \code{\link[=segments]{segments()}}  or \code{\link[=polypath]{polypath()}} as needed, in the usual
one-to-one or recycling way.

Graphical parameters are not able to be passed to the initial plot setup, but a plot
can be set up and then added to with this method.
}
