% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives_1D.r
\name{PRIMITIVE}
\alias{PRIMITIVE}
\alias{PRIMITIVE}
\alias{PRIMITIVE.PATH}
\alias{PRIMITIVE}
\alias{PRIMITIVE.sf}
\title{Generate a PRIMITIVES model.}
\usage{
PRIMITIVE(x, ...)

\method{PRIMITIVE}{PATH}(x, ...)

\method{PRIMITIVE}{sf}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments passed to methods}
}
\description{
A PRIMITIVES model is a decomposition of spatial forms to primitives. In the case
of POLYGON and MULTIPOLYGON and LINESTRING and MULTILINESTRING types this is the 
set of two-coordinate line segments.
}
\examples{

sf_obj <- inlandwaters[5, ]
sf_pslg <- PRIMITIVE(sf_obj)
plot(sf_pslg$vertex[, c("x_", "y_")], type = "n")
library(dplyr)
P1_segments <- function(x, ...) {
  s1 <- x$segment \%>\% 
      inner_join(x$vertex, c(".vertex0" = "vertex_")) \%>\% 
      select(x_, y_)
  s2 <- x$segment \%>\% 
      inner_join(x$vertex, c(".vertex1" = "vertex_")) \%>\% 
      select(x_, y_)
   graphics::segments(s1$x_, s1$y_, s2$x_, s2$y_, ...)
}
P1_segments(sf_pslg)
nc = sf::st_read(system.file("shape/nc.shp", package="sf"))
plot(sf::st_geometry(nc), border = "white")
nc_pslg <- PRIMITIVE(nc)
P1_segments(nc_pslg)

## find nodes?
nc_pslg$vertex 
}
