% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silicate-package.R
\name{sc-deprecated}
\alias{sc-deprecated}
\alias{sc-deprecated}
\alias{PRIMITIVE}
\title{Deprecated functions from silicate.}
\usage{
PRIMITIVE(x, ...)
}
\arguments{
\item{x}{input model}

\item{...}{arguments passed to methods}
}
\description{
`sc_uid(n = )` replaced by `sc_uid(x = )`
`PRIMITIVE()`
}
