% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.r, R/silicate-package.r
\name{sc-deprecated}
\alias{sc-deprecated}
\alias{sc_rand}
\alias{sc-deprecated}
\title{Deprecated functions from sc.}
\usage{
sc_rand(n = 1L)
}
\arguments{
\item{n}{see `sc_uid`}
}
\description{
Deprecated functions from sc.
}
