% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC0-model.R
\name{SC0}
\alias{SC0}
\alias{SC0.default}
\alias{SC0.pslg}
\title{Pure edge model, structural form}
\usage{
SC0(x, ...)

\method{SC0}{default}(x, ...)

\method{SC0}{pslg}(x, ...)
}
\arguments{
\item{x}{an object understood by silicate}

\item{...}{reserved for methods}
}
\value{
SC0 model with tables 'object' and 'vertex'
}
\description{
\code{SC0} is the simplest and most general of all silicate models. Composed of
an \code{object} and \code{vertex} table linked by nested vertex-index pairs.
}
\examples{
SC0(minimal_mesh)
SC0(minimal_mesh)
}
