% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_arc.R
\name{sc_arc}
\alias{sc_arc}
\alias{sc_arc}
\alias{sc_arc.default}
\alias{sc_arc}
\alias{sc_arc.ARC}
\title{Arcs for arc-node topology.}
\usage{
sc_arc(x, ...)

\method{sc_arc}{default}(x, ...)

\method{sc_arc}{ARC}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments for methods}
}
\description{
Return a label and vertex count of each arc.
}
\details{
Arcs are unbranched paths within the line segment graph. Nodes are the vertices where three or more arcs meet.

As with the `PATH` and `SC` models the arc id values will only be relevant when 
those entities are identified and labelled. Running `sc_arc` on a simple features model (for example) will
identify them and return a summary, but without having any record of what they refer to. Use `ARC(x)` first to 
work with models that don't included labels.
}
\examples{
sc_arc(minimal_mesh)
ARC(minimal_mesh)[["arc"]]
}
