% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_arc.R
\name{sc_arc}
\alias{sc_arc}
\alias{sc_arc}
\alias{sc_arc.default}
\alias{sc_arc}
\alias{sc_arc.PATH}
\title{Arcs for arc-node topology.}
\usage{
sc_arc(x, ...)

\method{sc_arc}{default}(x, ...)

\method{sc_arc}{PATH}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments for methods}
}
\description{
Arcs are unbranched paths within the line segment graph. Nodes are the vertices where three or more arcs meet.
}
