% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.r
\name{sc_coord}
\alias{sc_coord}
\alias{sc_coord}
\alias{sc_coord.PATH}
\title{Coordinate decomposition}
\usage{
sc_coord(x, ...)

\method{sc_coord}{PATH}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments passed to methods}
}
\description{
Collect all coordinates in one table, the path_link_vertex table
has the information about the original grouping.
}
\seealso{
`sc_path` for the central part of the model, `sc_object` for 
the features, and `PATH` for the full model.
}
