% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.r, R/sf-coord.r, R/sp-coord.R
\name{sc_coord}
\alias{sc_coord}
\alias{sc_coord}
\alias{sc_coord.PATH}
\alias{sc_coord}
\alias{sc_coord.sf}
\alias{sc_coord}
\alias{sc_coord.sfc}
\alias{sc_coord}
\alias{sc_coord.MULTIPOLYGON}
\alias{sc_coord}
\alias{sc_coord.POLYGON}
\alias{sc_coord}
\alias{sc_coord.MULTILINESTRING}
\alias{sc_coord}
\alias{sc_coord.LINESTRING}
\alias{sc_coord}
\alias{sc_coord.MULTIPOINT}
\alias{sc_coord}
\alias{sc_coord.POINT}
\alias{sc_coord}
\alias{sc_coord.Spatial}
\alias{sc_coord}
\alias{sc_coord.Polygons}
\alias{sc_coord}
\alias{sc_coord.Lines}
\alias{sc_coord}
\alias{sc_coord.default}
\title{Coordinate decomposition}
\usage{
sc_coord(x, ...)

\method{sc_coord}{PATH}(x, ...)

\method{sc_coord}{sf}(x, ...)

\method{sc_coord}{sfc}(x, ...)

\method{sc_coord}{MULTIPOLYGON}(x, ...)

\method{sc_coord}{POLYGON}(x, ...)

\method{sc_coord}{MULTILINESTRING}(x, ...)

\method{sc_coord}{LINESTRING}(x, ...)

\method{sc_coord}{MULTIPOINT}(x, ...)

\method{sc_coord}{POINT}(x, ...)

\method{sc_coord}{Spatial}(x, ...)

\method{sc_coord}{Polygons}(x, ...)

\method{sc_coord}{Lines}(x, ...)

\method{sc_coord}{default}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments passed to methods}
}
\description{
Collect all coordinates in one table, the path_link_vertex table
has the information about the original grouping.

Collect all coordinates in one table, the path_link_vertex table
has the information about the original grouping.
}
\examples{
data("sfzoo")
#sc_coord(sf::st_sfc(sfzoo))
# lapply(sfzoo, sc_coord)
#sc_coord(sf::st_sfc(sfzoo))
}
\seealso{
`sc_path` for the central part of the model, `sc_object` for 
the features, and `PATH` for the full model.

`sc_path` for the central part of the model, `sc_object` for 
the features, and `PATH` for the full model.
}
