% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_coord.R
\name{sc_coord}
\alias{sc_coord}
\alias{sc_coord.list}
\alias{sc_coord.default}
\alias{sc_coord.sfc_TIN}
\alias{sc_coord.matrix}
\alias{sc_coord.ARC}
\alias{sc_coord.PATH}
\alias{sc_coord.TRI}
\alias{sc_coord.PATH0}
\alias{sc_coord.SC0}
\alias{sc_coord.SC}
\alias{sc_coord.sf}
\alias{sc_coord.sfc}
\alias{sc_coord.sfc_GEOMETRYCOLLECTION}
\alias{sc_coord.MULTIPOLYGON}
\alias{sc_coord.POLYGON}
\alias{sc_coord.MULTILINESTRING}
\alias{sc_coord.LINESTRING}
\alias{sc_coord.MULTIPOINT}
\alias{sc_coord.POINT}
\alias{sc_coord.Spatial}
\alias{sc_coord.Polygons}
\alias{sc_coord.Lines}
\title{Coordinate decomposition}
\usage{
sc_coord(x, ...)

\method{sc_coord}{list}(x, ...)

\method{sc_coord}{default}(x, ...)

\method{sc_coord}{sfc_TIN}(x, ...)

\method{sc_coord}{matrix}(x, ...)

\method{sc_coord}{ARC}(x, ...)

\method{sc_coord}{PATH}(x, ...)

\method{sc_coord}{TRI}(x, ...)

\method{sc_coord}{PATH0}(x, ...)

\method{sc_coord}{SC0}(x, ...)

\method{sc_coord}{SC}(x, ...)

\method{sc_coord}{sf}(x, ...)

\method{sc_coord}{sfc}(x, ...)

\method{sc_coord}{sfc_GEOMETRYCOLLECTION}(x, ...)

\method{sc_coord}{MULTIPOLYGON}(x, ...)

\method{sc_coord}{POLYGON}(x, ...)

\method{sc_coord}{MULTILINESTRING}(x, ...)

\method{sc_coord}{LINESTRING}(x, ...)

\method{sc_coord}{MULTIPOINT}(x, ...)

\method{sc_coord}{POINT}(x, ...)

\method{sc_coord}{Spatial}(x, ...)

\method{sc_coord}{Polygons}(x, ...)

\method{sc_coord}{Lines}(x, ...)
}
\arguments{
\item{x}{input model}

\item{...}{arguments passed to methods}
}
\value{
data frame of all the coordinates in the order they occur
}
\description{
Collect all instances of all coordinates in one table. This complementary to the \code{sc_path} of
an object, since the number of coordinates per path gives a structural mapping into this set.

Collect all coordinates in one table, the path_link_vertex table
has the information about the original grouping.
}
\examples{
sc_coord(minimal_mesh)
sc_coord(SC(minimal_mesh))
}
\seealso{
\code{sc_path} for the central part of the model, \code{sc_object} for
the features, and \code{PATH} for the full model.

\code{sc_path} for the central part of the model, \code{sc_object} for
the features, and \code{PATH} for the full model.
}
