% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_edge.R
\name{sc_edge}
\alias{sc_edge}
\alias{sc_edge}
\alias{sc_edge.default}
\alias{sc_edge}
\alias{sc_edge.PATH}
\title{Unique edges for arc-node topology.}
\usage{
sc_edge(x, ...)

\method{sc_edge}{default}(x, ...)

\method{sc_edge}{PATH}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments for methods}
}
\description{
So-called "arcs" are unclosed paths that end in nodes, vertices shared by other arcs.
}
