% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_edge.R
\name{sc_edge}
\alias{sc_edge}
\alias{sc_edge}
\alias{sc_edge.default}
\alias{sc_edge}
\alias{sc_edge.PATH}
\title{Edges.}
\usage{
sc_edge(x, ...)

\method{sc_edge}{default}(x, ...)

\method{sc_edge}{PATH}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments for methods}
}
\description{
Simple binary relationships, a primitive composed of two vertices.
}
\details{
Edges are unique, undirected line segments. Compare to \code{sc_segment} which refers to all
instances of edges.
}
