% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGE.R
\name{sc_node}
\alias{sc_node}
\alias{sc_node}
\alias{sc_node.default}
\alias{sc_node}
\alias{sc_node.PATH}
\title{Nodes for arc-node topology.}
\usage{
sc_node(x, ...)

\method{sc_node}{default}(x, ...)

\method{sc_node}{PATH}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments for methods}
}
\description{
Nodes are the vertices in the graph that are shared by "arcs".
}
\seealso{
NARC
}
