% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_node.R
\name{sc_node}
\alias{sc_node}
\alias{sc_node.SC}
\alias{sc_node.SC0}
\alias{sc_node.default}
\alias{sc_node.PATH}
\alias{sc_node.ARC}
\title{Nodes for arc-node topology.}
\usage{
sc_node(x, ...)

\method{sc_node}{SC}(x, ...)

\method{sc_node}{SC0}(x, ...)

\method{sc_node}{default}(x, ...)

\method{sc_node}{PATH}(x, ...)

\method{sc_node}{ARC}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments for methods}
}
\value{
data frame of the nodes
}
\description{
Nodes are the vertices in the graph that are shared by "arcs".
}
\examples{
sc_node(ARC(minimal_mesh))
sc <- SC(routes)
library(dplyr)
plot(sc)
sc_node(sc) \%>\% inner_join(sc$vertex) \%>\% select(x_, y_) \%>\% points()
}
