% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.r
\name{sc_object}
\alias{sc_object}
\alias{sc_object.sf}
\title{Objects, features}
\usage{
sc_object(x, ...)

\method{sc_object}{sf}(x, ...)
}
\description{
The objects are the front end entities, the usual "GIS contract" objects, 
or features.
}
\examples{
example(sf::st_read)
sc_object(nc)
}
\seealso{
`sc_coord` for the coordinates part of the model, `sc_branch` for 
the central part of the model, and `BRANCH` for the full model.
}

