% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_object.R, R/TRI-model.R
\name{sc_object}
\alias{sc_object}
\alias{sc_object.default}
\alias{sc_object.sf}
\alias{sc_object.sfc}
\alias{sc_object.TRI}
\title{Objects, features}
\usage{
sc_object(x, ...)

\method{sc_object}{default}(x, ...)

\method{sc_object}{sf}(x, ...)

\method{sc_object}{sfc}(x, ...)

\method{sc_object}{TRI}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments passed to methods}
}
\value{
data frame of the object values
}
\description{
The objects are the front end entities, the usual "GIS contract" objects,
or features.

The objects are the front end entities, the usual "GIS contract" objects,
the features.
}
\examples{
sc_object(minimal_mesh)
sc_object(SC0(minimal_mesh))
}
\seealso{
\code{sc_coord} for the coordinates part of the model, \code{sc_path} for
the central part of the model, and \code{PATH} for the full model.

\code{sc_coord} for the coordinates part of the model, \code{sc_path} for
the central part of the model, and \code{PATH} for the full model.
}
