% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.r
\name{sc_path}
\alias{sc_path}
\title{Path decomposition}
\usage{
sc_path(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments passed to methods}
}
\description{
Start in the middle, and build the 'path-link-vertex' table.
}
\seealso{
`sc_coord` for the coordinates part of the model, `sc_object` for 
the features, and `PATH` for the full model.
}
