% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives_1D.r
\name{sc_primitive}
\alias{sc_primitive}
\title{Given a `PATH`` model decompose to 1-dimensional primitives.}
\usage{
sc_primitive(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments passed to methods}
}
\description{
Given a `PATH`` model decompose to 1-dimensional primitives.
}
