% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silicate-package.R
\docType{data}
\name{sfzoo}
\alias{sfzoo}
\alias{sfgc}
\title{Simple features zoo.}
\description{
Basic examples of each type of simple feature geometry. \code{sfzoo} is a list
with each of \emph{point}, \emph{multipoin}, \emph{linestring}, \emph{multilinestring}, \emph{polygon} and
\emph{multipolygon}. \code{sfgc} is a \emph{GEOMETRYCOLLECTION} of all the types in \code{sfzoo}.
}
\examples{
lapply(sfzoo, sc_coord)
lapply(sfzoo, sc_path)

## unsure how useful this is ...
sc_path(sfgc)
}
