% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRI-model.R
\name{TRI}
\alias{TRI}
\alias{TRI}
\alias{plot.TRI}
\title{Triangulation model}
\usage{
TRI(x, ...)

\method{plot}{TRI}(x, ...)
}
\arguments{
\item{x}{input model}

\item{...}{arguments passed to methods}
}
\value{
TRI model
}
\description{
Triangulation model
}
