% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRI0-model.R
\name{TRI0}
\alias{TRI0}
\alias{TRI0.PATH0}
\alias{TRI0.PATH}
\title{TRI0 model, structural triangulations}
\usage{
TRI0(x, ...)

\method{TRI0}{PATH0}(x, ...)

\method{TRI0}{PATH}(x, ...)
}
\arguments{
\item{x}{object understood by silicate (sf, sp, a silicate model, etc.)}

\item{...}{current unused}
}
\value{
TRI0 model
}
\description{
TRI0 model, structural triangulations
}
